feat: send not found logic to controller

+++ src/core/user/domain/service/user.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common'
+import { Inject, Injectable, UnauthorizedException } from '@nestjs/common'
 import { type CreateUserDTO } from '../dto/create-user.dto'
 import { type DeleteUserDTO } from '../dto/delete-user.dto'
 import { type FindUserDTO } from '../dto/find-user.dto'
@@ -15,7 +15,7 @@ export class UserUseCase {
     return await this.userRepository.createUser(user)
   }
 
-  async findUser (params: FindUserDTO): Promise<User> {
+  async findUser (params: FindUserDTO): Promise<User | null> {
     return await this.userRepository.findUser(params)
   }
 
@@ -28,6 +28,12 @@ export class UserUseCase {
   }
 
   async deleteUser (id: IDUserDTO, data: DeleteUserDTO): Promise<void> {
+    const user = await this.userRepository.findUserByID(id)
+
+    if (data.password === user?.password) {
+      throw new UnauthorizedException('Password') // change to custom message
+    }
+
     await this.userRepository.deleteUser(id, data)
   }
 }
